requirements.txt
flask
gunicorn
python-dotenv
resend
openpyxl
supabase